refactor: 회원가입 타입 분리
