+++ src/features/auth/types/store.ts
@@ -1,7 +1,5 @@
 /* eslint-disable unused-imports/no-unused-vars */
 
-import { Field } from '@/shared/constants/user';
-
 /**
  * Auth 관련 zustand actions
  */
@@ -49,13 +47,3 @@ export interface AuthState {
    */
   actions: AuthActions;
 }
-
-export type Category = { title: string; field: Field };
-
-export type AuthModalSize = 'md' | 'xl' | '5xl';
-
-export type SignUpSteps = 'signUp' | 'welcome' | 'categories' | 'startnow';
-
-export type SignUpConfig = {
-  [key in SignUpSteps]: { modalSize: AuthModalSize; contents: JSX.Element };
-};
